requirements.txt
playwright>=1.40.0
pyyaml>=6.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
